fix(tests): resolve all compilation errors in test suite
Fixed comprehensive test compilation issues across the codebase:

- Fixed tower dependency by adding missing 'util' feature
- Corrected ~30 test function signatures to properly return Result<()>
- Fixed P2PError variant case from MCP to Mcp
- Resolved adaptive module test compilation errors:
  - Replaced UserId::random() with manual hash generation
  - Fixed AdaptiveRouter::new calls to use new_with_id
  - Removed unused rand::RngCore import
- Cleaned up duplicate Ok(()) statements in test functions
- Fixed misplaced Ok(()) statements causing syntax errors

All tests in p2p-core (saorsa-core) now compile successfully.

+++ crates/p2p-core/src/identity/node_identity.rs
@@ -110,7 +110,7 @@ impl ProofOfWork {
             if *byte == 0 {
                 zero_bits += 8;
             } else {
-                zero_bits += byte.leading_zeros() as u32;
+                zero_bits += byte.leading_zeros();
                 break;
             }
         }
